\name{LungCancerACvsSCCGEO}
\alias{LungCancerACvsSCCGEO}
\alias{anoLC}
\alias{gsLC}



\title{Annotation of a small set of training and test set samples (30 total) used by team 221 in the IMPROVER DSC
 for the lung cancer sub-challenge.}

\description{
  The \code{LungCancerACvsSCCGEO} dataset consists: i) a data frame \code{anoLC} giving
  the file names of the affy cel files used in the training phase and their corresponding 
  phenotype (AC or SCC) and ii) \code{gsLC} the gold standard, i.e. the class membership
  of each test sample apearing in \code{anoLC}. 
 }

\usage{data(LungCancerACvsSCCGEO)}

\source{ GEO for the training data, while the test data comes from the citation below: 
Tarca AL, Lauria M, Unger M, Bilal E, Boue S, Dey KK, Hoeng J, Koeppl H,
Martin F, Meyer P, Nandy P, Norel R, Peitsch M, Rice JJ, Romero R,
Stolovitzky G, Talikka M, Xiang Y, Zechner C, and IMPROVER DSC Collaborators.
Strengths and limitations of microarray-based phenotype prediction: lessons
learned from the IMPROVER Diagnostic Signature Challenge.
\emph{Bioinformatics}. 2013;29(22):2892-2899.
doi:10.1093/bioinformatics/btt492. \cr
}

\keyword{datasets}


